qualifier/pyproject.toml
[tool.poetry]
name = "qualifier"
version = "0.1.0"
description = ""
authors = [
    "Josh S Wilkinson <JoshsWilkinson@Outlook.com>",
]
readme = "README.md"
packages = [
    { include = "images" },
]

[tool.poetry.dependencies]
python = "^3.11"
Pillow = "^10.0.0"
numpy = "^1.25.2"
opencv-python = "^4.8.0.76"


[build-system]
requires = [
    "poetry-core",
]
build-backend = "poetry.core.masonry.api"
